refactor(designer): commit関数をStoreへ移動
